requirements.txt
opencv-python
mediapipe
numpy
tensorflow-cpu
